[Port] new railmap from 1.12.2
